app/tests/dev_apis/test_notification_playground_api.py
from httpx import ASGITransport, AsyncClient
from starlette import status
from tortoise.contrib.test import TestCase

from app.main import app


class TestNotificationPlaygroundApi(TestCase):
    async def test_notification_playground_page_success(self):
        async with AsyncClient(transport=ASGITransport(app=app), base_url="http://test") as client:
            response = await client.get("/api/v1/dev/notifications-playground")
        assert response.status_code == status.HTTP_200_OK
        assert "text/html" in response.headers.get("content-type", "")
        assert "Notification API Playground" in response.text
